Move findOrThrow to shared Utils library
